+++ include/seqan3/io/sequence_file/output.hpp
@@ -64,12 +64,11 @@ namespace seqan3
  * FastA and FastQ, but some may also be interested in treating SAM or BAM files as sequence
  * files, discarding the alignment.
  *
- * The Sequence file abstraction supports writing four different fields:
+ * The Sequence file abstraction supports writing three different fields:
  *
  *   1. seqan3::field::seq
  *   2. seqan3::field::id
  *   3. seqan3::field::qual
- *   4. seqan3::field::seq_qual (sequence and qualities in one range)
  *
  * The member functions take any and either of these fields. If the field ID of an argument cannot be deduced, it
  * is assumed to correspond to the field ID of the respective template parameter.
@@ -111,10 +110,9 @@ namespace seqan3
  *
  * You may also use the output file's iterator for writing, however, this rarely provides an advantage.
  *
-* ### Writing record-wise (custom fields)
+ * ### Writing record-wise (custom fields)
  *
- * If you want to pass a combined object for SEQ and QUAL fields to push_back() / emplace_back(), or if you want
- * to change the order of the parameters, you can pass a non-empty fields trait object to the
+ * If you want to change the order of the parameters, you can pass a non-empty fields trait object to the
  * sequence_file_output constructor to select the fields that are used for interpreting the arguments.
  *
  * The following snippets demonstrates the usage of such a fields trait object.
@@ -182,7 +180,7 @@ public:
     //!\}
 
     //!\brief The subset of seqan3::field IDs that are valid for this file.
-    using field_ids            = fields<field::seq, field::id, field::qual, field::seq_qual>;
+    using field_ids            = fields<field::seq, field::id, field::qual, field::_seq_qual_deprecated>;
 
     static_assert([] () constexpr
                   {
@@ -196,7 +194,7 @@ public:
 
     static_assert([] () constexpr
                   {
-                      return !(selected_field_ids::contains(field::seq_qual) &&
+                      return !(selected_field_ids::contains(field::_seq_qual_deprecated) &&
                                (selected_field_ids::contains(field::seq) ||
                                (selected_field_ids::contains(field::qual))));
                   }(),
@@ -395,8 +393,7 @@ public:
         write_record(detail::get_or_ignore<field::seq>(r),
                      detail::get_or_ignore<field::id>(r),
                      detail::get_or_ignore<field::qual>(r),
-                     detail::get_or_ignore<field::seq_qual>(r));
-
+                     detail::get_or_ignore<field::_seq_qual_deprecated>(r));
     }
 
     /*!\brief           Write a record in form of a std::tuple to the file.
@@ -430,7 +427,7 @@ public:
         write_record(detail::get_or_ignore<selected_field_ids::index_of(field::seq)>(t),
                      detail::get_or_ignore<selected_field_ids::index_of(field::id)>(t),
                      detail::get_or_ignore<selected_field_ids::index_of(field::qual)>(t),
-                     detail::get_or_ignore<selected_field_ids::index_of(field::seq_qual)>(t));
+                     detail::get_or_ignore<selected_field_ids::index_of(field::_seq_qual_deprecated)>(t));
     }
 
     /*!\brief            Write a record to the file by passing individual fields.